refactor: organize command modules into subfolders
