+++ src/commands.ts
@@ -1,33 +1,33 @@
 import * as vscode from 'vscode';
 
 // Local imports - commands
-import { registerFileSelectionCommands } from './commands/fileSelectionCommands';
-import { registerLatexdiffCommands } from './commands/latexdiffCommands';
-import { registerGitCommands } from './commands/gitCommands';
+import { registerFileSelectionCommands } from '@commands/files/fileSelectionCommands';
+import { registerLatexdiffCommands } from '@commands/latex/latexdiffCommands';
+import { registerGitCommands } from '@commands/git/gitCommands';
 import { registerPackCommands } from './commands/packCommands';
 import { registerCleanCommands } from './commands/cleanCommands';
 import { registerMergeCommands } from './commands/mergeCommands';
 import { registerExecuteCommand } from './commands/executeCommand';
 import { TeXRAViewProvider } from './ViewProvider';
-import { registerLatexCommands } from './commands/latexCommands';
+import { registerLatexCommands } from '@commands/latex/latexCommands';
 import { registerImageCommands } from './commands/imageCommands';
 import { registerFigureCommands } from './commands/figCommands';
 import { registerTestCommands } from './commands/testCommands';
 import { registerXmlCommands } from './commands/xmlCommands';
 import { registerYamlCommands } from './commands/yamlCommands';
 import { registerAgentCommands } from './commands/agentCommands';
-import { registerApiKeyCommands } from './commands/apiKeyCommands';
+import { registerApiKeyCommands } from '@commands/api/apiKeyCommands';
 import { registerStateRestoreCommand } from './commands/stateRestoreCommand';
 import { registerTextEditorCommands } from './commands/textEditorCommands';
-import { registerLinterCommands } from './commands/linterCommands';
-import { registerWolframAlphaCommands } from './commands/wolframAlphaCommands';
-import { registerWolframScriptCommands } from './commands/wolframScriptCommands';
-import { registerHistoryCommands } from './commands/historyCommands';
+import { registerLinterCommands } from '@commands/latex/linterCommands';
+import { registerWolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
+import { registerWolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
+import { registerHistoryCommands } from '@commands/history/historyCommands';
 import { registerArXivCommands } from './commands/arXivCommands';
 import { registerCompareCommands } from './commands/compareCommands';
-import { registerProgressViewCommands } from './commands/progressViewCommands';
 import { registerHelpCommands } from './commands/helpCommands';
-import { registerOpenFileCommands } from './commands/openFileCommands';
+import { registerProgressViewCommands } from '@commands/progress/progressViewCommands';
+import { registerOpenFileCommands } from '@commands/files/openFileCommands';
 
 import * as logger from '@logger/logUtils';
 
@@ -86,26 +86,26 @@ export function registerCommands(context: vscode.ExtensionContext) {
 }
 
 // Add exports for the command modules
-export { fileSelectionCommands } from './commands/fileSelectionCommands';
-export { latexdiffCommands } from './commands/latexdiffCommands';
-export { gitCommands } from './commands/gitCommands';
+export { fileSelectionCommands } from '@commands/files/fileSelectionCommands';
+export { latexdiffCommands } from '@commands/latex/latexdiffCommands';
+export { gitCommands } from '@commands/git/gitCommands';
 export { packCommands } from './commands/packCommands';
 export { mergeCommands } from './commands/mergeCommands';
 export { executeCommand } from './commands/executeCommand';
-export { latexCommands } from './commands/latexCommands';
+export { latexCommands } from '@commands/latex/latexCommands';
 export { imageCommands } from './commands/imageCommands';
 export { figureCommands } from './commands/figCommands';
 export { testCommands } from './commands/testCommands';
 export { xmlCommands } from './commands/xmlCommands';
 export { yamlCommands } from './commands/yamlCommands';
-export { apiKeyCommands } from './commands/apiKeyCommands';
+export { apiKeyCommands } from '@commands/api/apiKeyCommands';
 export { stateRestoreCommand } from './commands/stateRestoreCommand';
 export { textEditorCommands } from './commands/textEditorCommands';
-export { linterCommands } from './commands/linterCommands';
-export { wolframAlphaCommands } from './commands/wolframAlphaCommands';
-export { wolframScriptCommands } from './commands/wolframScriptCommands';
-export { historyCommands } from './commands/historyCommands';
+export { linterCommands } from '@commands/latex/linterCommands';
+export { wolframAlphaCommands } from '@commands/wolfram/wolframAlphaCommands';
+export { wolframScriptCommands } from '@commands/wolfram/wolframScriptCommands';
+export { historyCommands } from '@commands/history/historyCommands';
 export { arXivCommands } from './commands/arXivCommands';
 export { compareCommands } from './commands/compareCommands';
 export { helpCommands } from './commands/helpCommands';
-export { registerOpenFileCommands as openFileCommands } from './commands/openFileCommands';
+export { registerOpenFileCommands as openFileCommands } from '@commands/files/openFileCommands';
