+++ src/commands/files/fileSelectionCommands.ts
