+++ src/commands/history/historyCommands.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { AgentHistoryViewProvider } from '../historyView/AgentHistoryViewProvider';
+import { AgentHistoryViewProvider } from '../../historyView/AgentHistoryViewProvider';
 
 export const historyCommands = {
   showHistory: 'texra.showAgentHistory',
