+++ src/commands/progress/progressViewCommands.ts
